Sentiment Analysis Python Notebooks
Two python notebooks with focus on trying to analyze a large corpus of movie reviews and derive the sentiment.

In the first part, we cover a wide variety of techniques for analyzing sentiment, which include the following.
- Unsupervised lexicon-based models
- Traditional supervised Machine Learning models

On the second We cover a wide variety of techniques of deep leaning for analyzing sentiment, which include the following.
- Newer supervised Deep Learning models
- Advanced supervised Deep Learning models

Besides looking at various approaches and models, we also focus on important aspects in the Machine Learning pipeline including text pre-processing, normalization, and in-depth analysis of models, including model interpretation and topic models. The key idea here is to understand how we tackle a problem like sentiment analysis on unstructured text, learn various techniques, models and understand how to interpret the results. This will enable you to use these methodologies in the future on your own datasets. Let's get started!